chore: preserve CodSpeed benchmark suites
