+++ Cargo.toml
@@ -126,11 +126,7 @@ name = "eviction"
 harness = false
 
 [[bench]]
-name = "eviction_revisions"
-harness = false
-
-[[bench]]
-name = "eviction_parallel"
+name = "eviction_walltime"
 harness = false
 required-features = ["rayon"]
